requirements.txt
discord.py[voice]==2.4.0
PyNaCl==1.5.0
feedparser==6.0.11
